+++ fortran/DarkEnergyFluid.f90
@@ -38,6 +38,7 @@
     procedure :: grho_de => TAxionEffectiveFluid_grho_de
     procedure :: PerturbedStressEnergy => TAxionEffectiveFluid_PerturbedStressEnergy
     procedure :: PerturbationEvolve => TAxionEffectiveFluid_PerturbationEvolve
+    procedure :: assume_scale_indep_lowz_growth => TAxionEffectiveFluid_assume_scale_indep_lowz_growth
     end type TAxionEffectiveFluid
 
     contains
@@ -213,7 +214,7 @@
             mu = 1/xc*(1-cos(this%theta_i))**((1-n)/2.)*sqrt((1-F)*(6*p+2)*this%theta_i/n/sin(this%theta_i))
             this%freq =  mu*(1-cos(this%theta_i))**((n-1)/2.)* &
                 sqrt(const_pi)*Gamma((n+1)/(2.*n))/Gamma(1+0.5/n)*2.**(-(n**2+1)/(2.*n)) &
-                    *3.**((1./n-1)/2)*this%a_c**(-6./(n+1)+3) &
+                *3.**((1./n-1)/2)*this%a_c**(-6./(n+1)+3) &
                 *( this%a_c**(6*n/(n+1.))+1)**(0.5*(1./n-1))
             this%n = n
         end if
@@ -296,5 +297,15 @@
     dgqe = ay(w_ix + 1) * grhov_t
 
     end subroutine TAxionEffectiveFluid_PerturbedStressEnergy
+
+    function TAxionEffectiveFluid_assume_scale_indep_lowz_growth(this)
+    !cs2 in TAxionEffectiveFluid_PerturbationEvolve is genuinely scale-dependent while the
+    !component oscillates, but that is confined to high redshift.
+    class(TAxionEffectiveFluid) :: this
+    logical :: TAxionEffectiveFluid_assume_scale_indep_lowz_growth
+
+    TAxionEffectiveFluid_assume_scale_indep_lowz_growth = this%zc > 50._dl
+
+    end function TAxionEffectiveFluid_assume_scale_indep_lowz_growth
 
     end module DarkEnergyFluid